Restructuration + Fix highlight Anchor + Update sidebar
J'ai restructuré le menu de nouveau et corrigé le problème de build sur le highlight (window undefined)
