+++ guides/FAQ/index.md
@@ -1,4 +1,4 @@
 ---
 title: FAQ
-order: 999999
+order: 999998
 ---
