+++ guides/Gameplay/index.md
@@ -1,0 +1,4 @@
+---
+title: Gameplay
+order: 0
+---
