--- guides/Guides/index.md
@@ -1,4 +1,0 @@
----
-title: 'Umamusume'
-order: 0
----
