+++ guides/Modes de jeu/Champion's Meeting/index.md
@@ -1,0 +1,4 @@
+---
+title: JcJ - Champion's Meeting
+order: 1
+---
