+++ guides/Modes de jeu/index.md
@@ -1,0 +1,4 @@
+---
+title: Modes de jeu
+order: 1
+---
