+++ package.json
@@ -4,7 +4,7 @@
   },
   "devDependencies": {
     "vitepress": "^1.6.3",
-    "vitepress-sidebar": "^1.31.1"
+    "vitepress-sidebar": "^1.33.0"
   },
   "scripts": {
     "dev": "vitepress dev",
